+++ backend/ApiRest/DTO/ProductoDTO.cs
@@ -2,10 +2,10 @@
 
 public class ProductoDTO
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string Nombre { get; set; } = null!;
     public decimal Precio { get; set; }
-    public int IdCat { get; set; }
+    public long IdCat { get; set; }
     public virtual CategoriaCreationDTO IdCatNavigation { get; set; }
     public virtual ICollection<ComandaInProductoDTO> Comanda { get; set; }
 }
