+++ backend/ApiRest/Repository/IMasterRepository.cs
@@ -4,8 +4,8 @@ public interface IMasterRepository<TEntity>
     where TEntity : class
 {
     Task<List<TEntity>> GetAll();
-    Task<TEntity?> GetById(int id);
+    Task<TEntity?> GetById(long id);
     Task<TEntity> Add(TEntity model);
     Task<TEntity> Update(TEntity model);
-    Task Delete(int id);
+    Task Delete(long id);
 }
